refactor: TimerPage의 상태관리를 해당 관심사 훅으로 분리
