+++ src/page/TimerPage/hooks/useNormalTimer.ts
@@ -67,3 +67,5 @@ export function useNormalTimer() {
     clearTimer,
   };
 }
+
+export type UseNormalTimerReturnType = ReturnType<typeof useNormalTimer>;
